# LiteWeightApiDocumentation/public/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "LiteWeight API",
    "version": "v1"
  },
  "paths": {
...
    "/shared-workouts/{sharedWorkoutId}": {
      "get": {
        "tags": [
          "SharedWorkouts"
        ],
        "summary": "Get Shared Workout",
        "description": "Gets a shared workout. Conditional that it was sent to the authenticated user.",
        "parameters": [
          {
            "name": "sharedWorkoutId",
            "in": "path",
            "description": "Id of the shared workout to get",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedWorkoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/shared-workouts/{sharedWorkoutId}/accept": {
      "post": {
        "tags": [
          "SharedWorkouts"
        ],
        "summary": "Accept Shared Workout",
        "description": "Accepts a shared workout and adds any exercises that the user doesn't already own to their owned exercises.<br /><br />\r\nAccepting a workout deletes the shared workout from the database and creates a workout with the values of that shared workout.<br/><br/><div style=\"background-color: #2f3031;padding: 12px 0px 12px 12px;border-left: 5px solid #F90258;\">Potential 400 Errors: <b>AlreadyExists, InvalidRequest, MaxLimit</b></div>",
        "parameters": [
          {
            "name": "sharedWorkoutId",
            "in": "path",
            "description": "Id of the shared workout to accept",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptSharedWorkoutRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptSharedWorkoutRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptSharedWorkoutRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptSharedWorkoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/shared-workouts/{sharedWorkoutId}/decline": {
      "delete": {
        "tags": [
          "SharedWorkouts"
        ],
        "summary": "Decline Shared Workout",
        "description": "Declines a workout and deletes it from the database. Conditional that the recipient matches the authenticated user.",
        "parameters": [
          {
            "name": "sharedWorkoutId",
            "in": "path",
            "description": "Id of the shared workout to decline",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/workouts/{workoutId}": {
      "get": {
        "tags": [
          "Workouts"
        ],
        "summary": "Get Workout",
        "description": "Gets a workout. Conditional that the creator matches the authenticated user.",
        "parameters": [
          {
            "name": "workoutId",
            "in": "path",
            "description": "Id of the workout to get",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkoutResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workouts"
        ],
        "summary": "Delete Workout",
        "description": "Deletes a given workout. Removes it from the authenticated user's list of workouts, and from the list of workouts on the exercises of the deleted workout.",
        "parameters": [
          {
            "name": "workoutId",
            "in": "path",
            "description": "Id of the workout to delete",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceNotFoundResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForbiddenResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptSharedWorkoutRequest": {
        "type": "object",
        "properties": {
          "workoutName": {
            "maxLength": 40,
            "type": "string",
            "description": "Optional name to set for the workout once accepted. Must be unique. If not specified, the created workout will have the name of the shared workout.",
            "nullable": true,
            "example": "Olympic Routine"
          }
        },
        "additionalProperties": false
      },
      "AcceptSharedWorkoutResponse": {
        "type": "object",
        "properties": {
          "newWorkoutInfo": {
            "$ref": "#/components/schemas/WorkoutInfoResponse"
          },
          "userExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedExerciseResponse"
            },
            "description": "New total list of exercises owned by the user as a result of accepting this workout. New exercises may have been added if the user did not already have them.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message describing the error.",
            "nullable": true,
            "example": "An example error message"
          },
          "errorType": {
            "type": "string",
            "description": "Specific type of error.",
            "nullable": true,
            "example": "InvalidRequest"
          },
          "requestErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ModelBindingError"
            },
            "description": "Lists any errors associated when binding the request.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ForbiddenResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Forbidden access"
          }
        },
        "additionalProperties": false
      },
...
      "ModelBindingError": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "description": "Property that is causing the error.",
            "nullable": true,
            "example": "name"
          },
          "message": {
            "type": "string",
            "description": "Error message for problematic property.",
            "nullable": true,
            "example": "The name property cannot be empty"
          }
        },
        "additionalProperties": false
      },
      "OwnedExerciseResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the exercise.",
            "nullable": true,
            "example": "88a54457-2253-404e-ac09-82a8f2ce5fb8"
          },
          "name": {
            "type": "string",
            "description": "Name of the exercise.",
            "nullable": true,
            "example": "Barbell Curl"
          },
          "defaultWeight": {
            "type": "number",
            "description": "Default weight of the exercise (lb). Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "double",
            "example": 65
          },
          "defaultSets": {
            "type": "integer",
            "description": "Default sets of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 3
          },
          "defaultReps": {
            "type": "integer",
            "description": "Default reps of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "format": "int32",
            "example": 15
          },
          "defaultDetails": {
            "type": "string",
            "description": "Default details of the exercise. Value that the exercise will be defaulted to when adding it to a workout.",
            "nullable": true,
            "example": "Don't overextend arms."
          },
          "videoUrl": {
            "type": "string",
            "description": "Video url of the exercise. Suggested use case is a video of how to perform the exercise.",
            "nullable": true,
            "example": "https://www.youtube.com/watch?v=kwG2ipFRgfo"
          },
          "focuses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of focuses of the exercise.",
            "nullable": true,
            "example": [
              "Biceps",
              "Forearms"
            ]
          },
          "workouts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedExerciseWorkoutResponse"
            },
            "description": "List of workouts this exercise is associated with.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OwnedExerciseWorkoutResponse": {
        "type": "object",
        "properties": {
          "workoutId": {
            "type": "string",
            "description": "Id of the workout.",
            "nullable": true,
            "example": "b7c3c321-d5a0-4448-978c-0a8b7709ceb2"
          },
          "workoutName": {
            "type": "string",
            "description": "Name of the workout.",
            "nullable": true,
            "example": "Main Workout"
          }
        },
        "additionalProperties": false
      },
...
      "ResourceNotFoundResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The workout with id=cfbb1fd4-6d29-465c-9ea0-4938407ac65b was not found"
          }
        },
        "additionalProperties": false
      },
...
      "RoutineDayResponse": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "Arbitrary tag of the day.",
            "nullable": true,
            "example": "Back and Biceps Day"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineExerciseResponse"
            },
            "description": "List of exercises for the given day.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoutineExerciseResponse": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string",
            "description": "Id of the exercise (reference to the list of exercises on the user).",
            "nullable": true,
            "example": "88a54457-2253-404e-ac09-82a8f2ce5fb8"
          },
          "completed": {
            "type": "boolean",
            "description": "Has the user completed this exercise?"
          },
          "weight": {
            "maximum": 99999,
            "minimum": 0,
            "type": "number",
            "description": "Weight of the exercise (lb).",
            "format": "double",
            "example": 30
          },
          "sets": {
            "maximum": 99,
            "minimum": 0,
            "type": "integer",
            "description": "Number of sets for the exercise.",
            "format": "int32",
            "example": 3
          },
          "reps": {
            "maximum": 999,
            "minimum": 0,
            "type": "integer",
            "description": "Number of reps for the exercise.",
            "format": "int32",
            "example": 15
          },
          "details": {
            "maxLength": 120,
            "type": "string",
            "description": "Details of the exercise.",
            "nullable": true,
            "example": "Don't overextend arms."
          }
        },
        "additionalProperties": false
      },
      "RoutineResponse": {
        "type": "object",
        "properties": {
          "weeks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineWeekResponse"
            },
            "description": "List of weeks in the routine.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RoutineWeekResponse": {
        "type": "object",
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoutineDayResponse"
            },
            "description": "List of days in the routine.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ServerErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "The server has encountered an error"
          }
        },
        "additionalProperties": false
      },
...
      "SharedDayResponse": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "description": "Arbitrary tag of the day.",
            "nullable": true,
            "example": "Cardio and Legs Day"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedExerciseResponse"
            },
            "description": "List of exercises for the given day.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SharedExerciseResponse": {
        "type": "object",
        "properties": {
          "exerciseName": {
            "type": "string",
            "description": "Name of the exercise.",
            "nullable": true,
            "example": "Squat"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the exercise (lb).",
            "format": "double",
            "example": 215
          },
          "sets": {
            "type": "integer",
            "description": "Number of sets for the exercise.",
            "format": "int32",
            "example": 3
          },
          "reps": {
            "type": "integer",
            "description": "Number of reps for the exercise.",
            "format": "int32",
            "example": 15
          },
          "details": {
            "type": "string",
            "description": "Details of the exercise.",
            "nullable": true,
            "example": "Don't lock knees."
          }
        },
        "additionalProperties": false
      },
      "SharedRoutineResponse": {
        "type": "object",
        "properties": {
          "weeks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedWeekResponse"
            },
            "description": "List of weeks in the routine.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SharedWeekResponse": {
        "type": "object",
        "properties": {
          "days": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedDayResponse"
            },
            "description": "List of days in the routine.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SharedWorkoutDistinctExerciseResponse": {
        "type": "object",
        "properties": {
          "exerciseName": {
            "type": "string",
            "description": "Name of the exercise.",
            "nullable": true,
            "example": "Squat"
          },
          "videoUrl": {
            "type": "string",
            "description": "Video url of the exercise.",
            "nullable": true,
            "example": "https://www.youtube.com/watch?v=Dy28eq2PjcM"
          },
          "focuses": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "List of focuses of the exercise.",
            "nullable": true,
            "example": [
              "Legs",
              "Strength Training"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "SharedWorkoutResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the shared workout.",
            "nullable": true,
            "example": "3ac84a61-4822-4ba3-ac93-626fdf087acf"
          },
          "workoutName": {
            "type": "string",
            "description": "Name of the shared workout.",
            "nullable": true,
            "example": "Legs Galore"
          },
          "senderId": {
            "type": "string",
            "description": "Id of the user who shared the workout.",
            "nullable": true,
            "example": "juo06et3-n81k-9bb1-61dj-j12k1152hae1"
          },
          "senderUsername": {
            "type": "string",
            "description": "Username of the user who shared the workout.",
            "nullable": true,
            "example": "arthur_v"
          },
          "recipientId": {
            "type": "string",
            "description": "Id of the recipient of the shared workout.",
            "nullable": true,
            "example": "f1e03cd1-e62c-4a53-84ed-498c72776fc2"
          },
          "routine": {
            "$ref": "#/components/schemas/SharedRoutineResponse"
          },
          "distinctExercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SharedWorkoutDistinctExerciseResponse"
            },
            "description": "List of distinct exercises in the routine of the shared workout.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UnauthorizedResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Response message.",
            "nullable": true,
            "example": "Unauthorized access"
          }
        },
        "additionalProperties": false
      },
...
      "WorkoutInfoResponse": {
        "type": "object",
        "properties": {
          "workoutId": {
            "type": "string",
            "description": "Id of the workout.",
            "nullable": true,
            "example": "84a34611-9f4c-443f-97f9-cbf5cde69c65"
          },
          "workoutName": {
            "type": "string",
            "description": "Name of the workout.",
            "nullable": true,
            "example": "3-Day Split"
          },
          "currentWeek": {
            "type": "integer",
            "description": "Index of the current week the user is on.",
            "format": "int32",
            "example": 2
          },
          "currentDay": {
            "type": "integer",
            "description": "Index of the current day of the current week the user is on.",
            "format": "int32",
            "example": 0
          },
          "lastSetAsCurrentUtc": {
            "type": "string",
            "description": "Timestamp of when the workout was last set as the current workout (UTC).",
            "nullable": true,
            "example": "2023-04-06T23:20:39.665047Z"
          },
          "timesRestarted": {
            "type": "integer",
            "description": "Total times the workout has been restarted.",
            "format": "int32",
            "example": 15
          },
          "averageExercisesCompleted": {
            "type": "number",
            "description": "Average of all exercises completed on the workout.",
            "format": "double",
            "example": 94.2
          }
        },
        "additionalProperties": false
      },
      "WorkoutResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the workout.",
            "nullable": true,
            "example": "b7c3c321-d5a0-4448-978c-0a8b7709ceb2"
          },
          "name": {
            "type": "string",
            "description": "Name of the workout.",
            "nullable": true,
            "example": "Main Workout"
          },
          "creationUtc": {
            "type": "string",
            "description": "Timestamp of when the workout was created (UTC).",
            "nullable": true,
            "example": "2023-04-23T16:49:02.310661Z"
          },
          "creatorId": {
            "type": "string",
            "description": "Id of the user who created the workout.",
            "nullable": true,
            "example": "a7be7348-bca5-466c-b290-55ae38e2bad0"
          },
          "routine": {
            "$ref": "#/components/schemas/RoutineResponse"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "BearerDefinition": {
        "type": "http",
        "description": "Token authentication. \n\n 'Bearer TOKEN'\n\nBearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}